feat: cjs enter of chatkit types
