+++ .changeset/chatkit-types-runtime-constants.md
@@ -1,0 +1,5 @@
+---
+"@xpert-ai/chatkit-types": patch
+---
+
+Expose runtime constants from a dedicated `./constants` subpath with ESM and CJS entry points.
