+++ packages/chatkit/src/index.ts
@@ -1,3 +1,4 @@
+export * from './constants.js';
 export * from './interrupt.js';
 export * from './message.js';
 export * from './options.js';
